+++ turkdemy/urls.py
@@ -13,11 +13,11 @@ urlpatterns: list[URLPattern | URLResolver] = [
     path("i18n/", include("django.conf.urls.i18n")),
     path("health/", include("apps.health.urls")),
     path("api/v1/", include("apps.api.urls")),
+    path("accounts/", include("allauth.urls")),
 ]
 
 urlpatterns += i18n_patterns(
     path("admin/", admin.site.urls),
-    path("accounts/", include("django.contrib.auth.urls")),
     path("", include("apps.public.urls")),
     path("", include("apps.leads.urls")),
 )